requirements.txt
streamlit
PyMuPDF
python-pptx
transformers
torch